python/poc-langchain/requirements.txt
openai==1.61.0
langchain==0.3.17
langchain-openai==0.3.3
python-dotenv==1.0.1
faiss-cpu==1.8.0
pypdf==4.1.0